Moving TemplateExpressionError definition out of numvalue.py
