+++ pyomo/core/expr/numvalue.py
@@ -23,14 +23,16 @@ from pyomo.core.expr.expr_common import \
      _iadd, _isub, _imul, _idiv,
      _ipow, _lt, _le, _eq)
 
+from pyomo.core.expr.expr_errors import TemplateExpressionError
+
 logger = logging.getLogger('pyomo.core')
 
 
-class TemplateExpressionError(ValueError):
+# class TemplateExpressionError(ValueError):
 
-    def __init__(self, template, *args, **kwds):
-        self.template = template
-        super(TemplateExpressionError, self).__init__(*args, **kwds)
+#     def __init__(self, template, *args, **kwds):
+#         self.template = template
+#         super(TemplateExpressionError, self).__init__(*args, **kwds)
 
 def _generate_sum_expression(etype, _self, _other):
     raise RuntimeError("incomplete import of Pyomo expression system")  #pragma: no cover
